moved generic math function to common module
